loading and rendering sprites update
i also added playerstates

+++ Minigin/Font.cpp
@@ -8,7 +8,7 @@ TTF_Font* dae::Font::GetFont() const {
 
 dae::Font::Font(const std::string& fullPath, unsigned int size) : m_font(nullptr)
 {
-	m_font = TTF_OpenFont(fullPath.c_str(), size);
+	m_font = TTF_OpenFont(fullPath.c_str(), static_cast<int>(size));
 	if (m_font == nullptr) 
 	{
 		throw std::runtime_error(std::string("Failed to load font: ") + SDL_GetError());
